+++ include/threadpool/shutdown/wait_active.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <memory>
 
-namespace thread_pool
+namespace threadpool
 {
 	template <typename _Pool>
 	struct wait_for_active_task
